Add submodule support in defs
