+++ crates/parser/src/db.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use diagnostics::{Diagnostics, WithDiagnostics};
 use diagnostics_proc_macros::with_diagnostics;
 use filesystem::db::FilesGroup;
-use filesystem::ids::{FileId, ModuleId};
+use filesystem::ids::FileId;
 use syntax::node::ast::SyntaxFile;
 use syntax::node::db::{AsSyntaxGroup, SyntaxGroup};
 
@@ -20,10 +20,6 @@ pub trait ParserGroup: SyntaxGroup + AsSyntaxGroup + FilesGroup {
         &self,
         file_id: FileId,
     ) -> WithDiagnostics<Option<Arc<SyntaxFile>>, ParserDiagnostic>;
-    fn module_syntax(
-        &self,
-        module_id: ModuleId,
-    ) -> WithDiagnostics<Option<Arc<SyntaxFile>>, ParserDiagnostic>;
 }
 
 #[with_diagnostics]
@@ -35,13 +31,4 @@ pub fn file_syntax(
     let s = db.file_content(file_id)?;
     let parser = Parser::from_text(db.as_syntax_group(), file_id, s.as_str());
     Some(Arc::new(parser.parse_syntax_file().propagte(diagnostics)))
-}
-
-#[with_diagnostics]
-pub fn module_syntax(
-    diagnostics: &mut Diagnostics<ParserDiagnostic>,
-    db: &dyn ParserGroup,
-    module_id: ModuleId,
-) -> Option<Arc<SyntaxFile>> {
-    db.file_syntax(db.module_file(module_id)?).propagte(diagnostics)
 }
