+++ crates/parser/test_data/cairo_files/test1.cairo
@@ -1,3 +1,6 @@
+mod submod;
+use submod::something;
+
 // Func foo
 // Second line
 ;
